requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
youtube-transcript-api==1.2.4
google-genai
python-dotenv==1.0.1
